OPOA Framework, OPOA standards setters. V2.4.1
+ EasyOPOA.preFirstHash function support in the browser before the first
request and resolve window.location.hash perform this function returns
false to prevent EasyOPOA default resolution. Usually when the browser
loads, to override the default of Hash resolution
+ Action List name hash EasyOPOA.loadLinked function supports manually
in order to load the specified. In turn can be used to achieve the
effect of multi-level hit loaded. For example, loading a hash request to
start within the hash request, we turn to load
+ start function supports direct incoming opoa single instance, you can
not use an array
# Cancel object cache, the cache object to prevent internal causes of
invalidity reload
# When calling addActionMap custom ActionMaps, in clearOPOA stored hash
of loading and urlErrors